Working on the main page. Added current weather situation about place that user picked

+++ lib/presentation/models/openweather_model/weather_model.dart
@@ -10,12 +10,12 @@ class WeatherModel{
 
 
   @JsonKey(name: "dt")
-  double? currentTime;
+  int? currentTime;
 
   @JsonKey(name: "sunrise")
-  double? sunrise;
+  int? sunrise;
   @JsonKey(name: "sunset")
-  double? sunset;
+  int? sunset;
 
 
   @JsonKey(name: "feels_like")
